Moved DependencyInfo struct and related code to a package
