+++ pkg/dependencies/dependencyinfo.go
@@ -1,11 +1,11 @@
-package main
+package dependencies
 
 type DependencyInfo struct {
-	Dependencies []dependency      `json:"dependencies"`
+	Dependencies []Dependency      `json:"dependencies"`
 	Licenses     map[string]string `json:"licenseInfo"`
 }
 
-type dependency struct {
+type Dependency struct {
 	Name     string   `json:"name"`
 	Version  string   `json:"version"`
 	Licenses []string `json:"licenses"`
@@ -13,7 +13,7 @@ type dependency struct {
 
 func NewDependencyInfo() DependencyInfo {
 	return DependencyInfo{
-		Dependencies: []dependency{},
+		Dependencies: []Dependency{},
 		Licenses:     map[string]string{},
 	}
 }
